aging/requirements.txt
astroid==2.4.2
autopep8==1.5.4
certifi==2020.12.5
chardet==4.0.0
Deprecated==1.2.11
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.6.0
PyGithub==1.54.1
PyJWT==1.7.1
pylint==2.6.0
requests==2.25.1
six==1.15.0
toml==0.10.2
urllib3==1.26.3
wrapt==1.12.1
